Inicio tests producto matriz

+++ MatlabUtils.h
@@ -6,7 +6,7 @@
 double norm(double *v);
 double dot(double *v1, double *v2);
 double sign(double n);
-double **zeros(size_t rows, size_t cols);
+double **zeros(unsigned int rows, unsigned int cols);
 double det(double M[][3]);
 double det2x2(double m[][2]);
 int roots(double *coef, double **sols_reales);
